+++ packages/plugin-rax-pha/src/manifestHelpers.js
@@ -24,13 +24,37 @@ const retainKeys = [
   'links',
   'scripts',
   'offlineResources',
+  'packageResources',
   'manifestPrefetchExpires',
   'manifestPrefetchMaxAge',
+  'maxAge',
+  'expires',
   'queryParamsPassKeys',
   'queryParamsPassIgnoreKeys',
   'splashViewTimeout',
+  'splashViewAutoClose',
   'swiperThreshold',
   'requestHeaders',
+  'enablePoplayer',
+  'disableCapture',
+  'enablePullRefresh',
+  'pullRefreshBackgroundColor',
+  'pullRefreshColorScheme',
+  'pullRefresh',
+  'cacheQueryParams',
+];
+
+// do not decamelize list
+const camelizeKeys = [
+  'appKey',
+  'dataType',
+  'valueType',
+  'isSec',
+  'LoginRequest',
+  'sessionOption',
+  'AntiCreep',
+  'AntiFlood',
+  'needLogin',
 ];
 
 // transform app config to decamelize
@@ -51,7 +75,11 @@ function transformAppConfig(appConfig, isRoot = true, parentKey) {
     if (key === 'pageHeader') {
       key = 'tabHeader';
     }
-    const transformKey = decamelize(key);
+
+    let transformKey = key;
+    if (camelizeKeys.indexOf(key) === -1) {
+      transformKey = decamelize(key);
+    }
     if (key === 'window') {
       Object.assign(data, transformAppConfig(value, false));
     } else if (typeof value === 'string' || typeof value === 'number') {
@@ -63,15 +91,23 @@ function transformAppConfig(appConfig, isRoot = true, parentKey) {
           delete item.text;
         }
         if (typeof item === 'object') {
-          if (key === 'dataPrefetch' && !item.header) {
+          if (key === 'dataPrefetch') {
             // hack: No header will crash in Android
-            item.header = {};
+            if (!item.header) {
+              item.header = {};
+            }
+
+            // no prefetchKey will crash in Android TaoBao 9.26.0
+            if (!item.prefetchKey) {
+              item.prefetchKey = 'mtop';
+            }
           }
           return transformAppConfig(item, false, key);
         }
         return item;
       });
-    } else if (key === 'requestHeaders') { // keys of requestHeaders should not be transformed
+    } else if (key === 'requestHeaders') {
+      // keys of requestHeaders should not be transformed
       data[transformKey] = value;
     } else if (typeof value === 'object' && !(parentKey === 'dataPrefetch' && (key === 'header' || key === 'data'))) {
       data[transformKey] = transformAppConfig(value, false, key);
@@ -102,6 +138,7 @@ function getRealPageInfo({ urlPrefix, urlSuffix = '' }, page) {
   }
 
   delete page.source;
+
   return {
     pageUrl,
     entryName,
@@ -117,25 +154,37 @@ function changePageInfo({ urlPrefix, urlSuffix = '', cdnPrefix, isTemplate, inli
   if (!source && !name) {
     return page;
   }
+
   const { document, custom } = applyMethod('rax.getDocument', { name, source }) || {};
-  const { entryName, pageUrl } = getRealPageInfo({
-    urlPrefix,
-    urlSuffix,
-  }, page);
+  const { entryName, pageUrl } = getRealPageInfo(
+    {
+      urlPrefix,
+      urlSuffix,
+    },
+    page,
+  );
+
   if (entryName) {
+    if (page.url) {
+      page.path = page.url;
+      delete page.url;
+      return page;
+    }
+
     if (!page.path || !page.path.startsWith('http')) {
       page.path = pageUrl;
     }
 
-    if (isTemplate) {
+    // template and no frames under the page
+    if (isTemplate && !Array.isArray(page.frames)) {
       if (custom) {
         page.document = document;
-      } else {
-        // add script and stylesheet
-        page.script = `${cdnPrefix + entryName}.js`;
-        if (!inlineStyle) {
-          page.stylesheet = `${cdnPrefix + entryName}.css`;
-        }
+        return page;
+      }
+      // add script and stylesheet
+      page.script = `${cdnPrefix + entryName}.js`;
+      if (!inlineStyle || (typeof inlineStyle === 'object' && inlineStyle.forceEnableCSS)) {
+        page.stylesheet = `${cdnPrefix + entryName}.css`;
       }
     }
   }
@@ -147,18 +196,39 @@ function changePageInfo({ urlPrefix, urlSuffix = '', cdnPrefix, isTemplate, inli
  * set real url to manifest
  */
 function setRealUrlToManifest(options, manifest) {
-  const { urlPrefix, cdnPrefix } = options;
+  const { urlPrefix, cdnPrefix, api } = options;
+  const { applyMethod } = api;
   if (!urlPrefix) {
     return manifest;
   }
 
   const { app_worker, tab_bar, pages } = manifest;
-  if (app_worker && app_worker.url) {
+  if (app_worker && app_worker.url && !app_worker.url.startsWith('http')) {
     app_worker.url = cdnPrefix + app_worker.url;
   }
 
-  if (tab_bar && tab_bar.source && !tab_bar.url) {
-    tab_bar.url = getRealPageInfo(options, tab_bar).pageUrl;
+  if (tab_bar && tab_bar.source) {
+    const { document, custom } = applyMethod('rax.getDocument', { name: tab_bar.name, source: tab_bar.source }) || {};
+    if (!tab_bar.url) {
+      if (custom) {
+        tab_bar.html = document;
+      }
+      // TODO: iOS issue
+      // TODO: should remove it in PHA 2.x
+      // PHA 1.x should inject `url` to be a base url to load assets
+      tab_bar.url = getRealPageInfo(options, tab_bar).pageUrl;
+      // TODO: Android issue
+      // TODO: should remove it in PHA 2.x
+      // same as iOS issue
+      tab_bar.name = new URL(tab_bar.url).origin;
+    }
+    delete tab_bar.source;
+  }
+
+  // items is `undefined` will crash in PHA
+  if (tab_bar && tab_bar.list) {
+    tab_bar.items = tab_bar.list.map(() => ({}));
+    delete tab_bar.list;
   }
 
   if (pages && pages.length > 0) {
@@ -171,7 +241,22 @@ function setRealUrlToManifest(options, manifest) {
       }
 
       if (page.tab_header && page.tab_header.source) {
-        page.tab_header.url = getRealPageInfo(options, page.tab_header).pageUrl;
+        const { document, custom } =
+          applyMethod('rax.getDocument', { name: page.tab_header.name, source: page.tab_header.source }) || {};
+        if (!page.tab_header.url) {
+          if (custom) {
+            page.tab_header.html = document;
+          }
+          // TODO: iOS issue
+          // TODO: should remove it in PHA 2.x
+          // PHA 1.x should inject `url` to be a base url to load assets
+          page.tab_header.url = getRealPageInfo(options, page.tab_header).pageUrl;
+          // TODO: Android issue
+          // TODO: should remove it in PHA 2.x
+          // same as iOS issue
+          page.tab_header.name = new URL(page.tab_header.url).origin;
+        }
+        delete page.tab_header.source;
       }
       return changePageInfo(options, page, manifest);
     });