+++ packages/plugin-rax-store/src/utils/getPath.ts
@@ -13,26 +13,11 @@ export function getAppStorePath(srcPath: string) {
   return storeFileType ? path.join(srcPath, `store${storeFileType}`) : '';
 }
 
-function getPagePath(srcPath: string, pageName: string) {
-  return path.join(srcPath, 'pages', pageName);
-}
-
-export function getPageStorePath(srcPath: string, pageName: string) {
-  const pageNameDir = getPagePath(srcPath, pageName);
+export function getPageStorePath(pageComponentPath: string) {
+  const pageNameDir = path.dirname(pageComponentPath);
   const storeFileType = getStoreFileType(pageNameDir);
-  // e.g: src/pages/Home/store.ts
+  // e.g: src/pages/Home/store.ts or src/pages/Home/About/store.ts
   return storeFileType ? path.join(pageNameDir, `store${storeFileType}`) : '';
-}
-
-export function getRaxPagesName(rootDir: string) {
-  const pagesPath = getRaxPagesPath(rootDir);
-  return pagesPath.map(getRaxPageName);
-}
-
-export function getRaxPageName(pagePath: string) {
-  const dir = path.dirname(pagePath);
-  const pageName = path.parse(dir).name;
-  return pageName;
 }
 
 export function getRaxPagesPath(rootDir: string) {