+++ packages/plugin-rax-web/src/setLocalBuilder.ts
@@ -119,7 +119,7 @@ export default (api, documentPath?: string | undefined) => {
 };
 
 function addReloadByDocumentChange(rootDir, entries) {
-  const watcher = chokidar.watch(`${rootDir}/src/document/index.@(tsx|js?(x))`, {
+  const watcher = chokidar.watch(`${rootDir}/src/document/**`, {
     ignoreInitial: true,
     atomic: 300,
   });