+++ packages/plugin-ssr/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v1.3.12
+
+- Feat: escape data to prevent xss
+
+## v1.3.11
+
+- Feat: support sdk combo
+
 ## v1.3.10
 
 - Fix: pass context when generate html
@@ -11,7 +19,7 @@
 
 ## v1.3.8
 
-- Feat: support query is `csr=true`  with request path end is `.html`, devServer will return CSR result
+- Feat: support query is `csr=true` with request path end is `.html`, devServer will return CSR result
 
 ## v1.3.7
 
@@ -46,4 +54,3 @@
 ## v1.3.0
 
 - Feat: support config `hash` in build.json
-